feat: make the pytest e2e harness a reusable plugin
Package the harness as installable `opensovd-e2e` so any OpenSOVD repo can
drive its own binary with the same fixtures instead of duplicating the setup.
The in-repo suite keeps its behaviour, now layering only its specifics on top.

Issue: eclipse-opensovd/opensovd-core#80
