+++ pyproject.toml
@@ -24,7 +24,7 @@ dev = [
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
-pythonpath = ["tests"]
+pythonpath = ["tests", "opensovd-e2e/src"]
 # importlib mode lets per-crate dirs (opensovd-gateway/cli/, opensovd-mcp/cli/)
 # share filenames like test_version.py without basename collisions.
 addopts = "--import-mode=importlib"
@@ -32,7 +32,7 @@ markers = [
     "conformance: Conformance tests (require bru CLI)",
     "schema: Tests that validate response schema",
     "bruno: Bruno API conformance tests",
-    "req: Requirement ID for traceability (e.g., @pytest.mark.req('SOVD-7.1.1'))",
+    # "req" marker is registered by the opensovd_e2e plugin.
 ]
 
 [tool.ruff]
@@ -60,4 +60,4 @@ select = [
 include = ["tests"]
 
 [tool.ty.environment]
-extra-paths = ["tests"]
+extra-paths = ["tests", "opensovd-e2e/src"]
